float conversion fix

+++ federated.cpp
@@ -91,7 +91,7 @@ bool shareW_out(uint16_t* nBatchesOnDevice) {
 		for(size_t j = 0; j < RESERVOIR_SIZE; j++) {
 			old_weight = W_a.weights[i][j];
 			
-			W_a.weights[i][j] = (W_a.weights[i][j] * n_a + W_b.weights[i][j] * n_b) / n_tot;
+			W_a.weights[i][j] = (W_a.weights[i][j] * (float) n_a + W_b.weights[i][j] * (float) n_b) / (float) n_tot;
 			new_weight = W_a.weights[i][j];
 
 			Serial.print("i: ");
